Restore the previous mechanism of handling simple_identity_encoding
The change was originally made with the intent to allow
simple_identity_encoding to work with PDF forms and to a secondary
extent make the fonts embedded in the PDF more correct. In the end the
custom encoding with embedded CMap file in the PDF also did not work
with PDF forms and the new method caused issues when attempting to edit the
PDF in Adobe Acrobat. So this commit reverts back to that earlier approach.